Move rules to separate project
