+++ Runner/Runner.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.0</TargetFramework>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
@@ -27,11 +27,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\ValidationLibrary\ValidationLibrary.csproj" />
     <ProjectReference Include="..\ValidationLibrary.Slack\ValidationLibrary.Slack.csproj" />
     <ProjectReference Include="..\ValidationLibrary.GitHub\ValidationLibrary.GitHub.csproj" />
     <ProjectReference Include="..\ValidationLibrary.Csv\ValidationLibrary.Csv.csproj" />
     <ProjectReference Include="..\ValidationLibrary.MarkdownGenerator\ValidationLibrary.MarkdownGenerator.csproj" />
+    <ProjectReference Include="..\ValidationLibrary.Rules\ValidationLibrary.Rules.csproj" />
   </ItemGroup>
 
 </Project>
