+++ ValidationLibrary.Rules.Tests/HasLicenseRuleTests.cs
