+++ ValidationLibrary.Rules/HasDescriptionRule.cs
