+++ ValidationLibrary.Rules/HasLicenseRule.cs
