+++ ValidationLibrary.Rules/HasNotManyStaleBranchesRule.cs
