+++ ValidationLibrary.Rules/ValidationLibrary.Rules.ruleset
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<RuleSet Name="Microsoft Managed Recommended Rules" Description="These rules focus on the most critical problems in your code, including potential security holes, application crashes, and other important logic and design errors. It is recommended to include this rule set in any custom rule set you create for your projects." ToolsVersion="10.0">
+  <Localization ResourceAssembly="Microsoft.VisualStudio.CodeAnalysis.RuleSets.Strings.dll" ResourceBaseName="Microsoft.VisualStudio.CodeAnalysis.RuleSets.Strings.Localized">
+    <Name Resource="MinimumRecommendedRules_Name" />
+    <Description Resource="MinimumRecommendedRules_Description" />
+  </Localization>
+  <Rules AnalyzerId="Microsoft.Analyzers.ManagedCodeAnalysis" RuleNamespace="Microsoft.Rules.Managed">
+    <Rule Id="CA1303" Action="Hidden" />
+    <Rule Id="CA2208" Action="Hidden" />
+    <Rule Id="CA1016" Action="Hidden" />
+    <Rule Id="CA1056" Action="Hidden" />
+    <Rule Id="CA1819" Action="Hidden" />
+    <Rule Id="CS1591" Action="Hidden" />
+  </Rules>
+</RuleSet>
