+++ ValidationLibrary/RepositoryValidator.cs
@@ -3,7 +3,6 @@ using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Octokit;
-using ValidationLibrary.Rules;
 
 namespace ValidationLibrary
 {
